refactor: extract extension source metadata lifecycle
